requirements.txt
streamlit>=1.44.0,<2.0.0
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-openai>=0.3.0,<0.4.0
langchain-chroma>=0.2.0,<0.3.0
langchain-huggingface>=0.3.0,<0.4.0
langchain-text-splitters>=0.3.0,<0.4.0
chromadb>=0.5.0,<1.0.0
sentence-transformers>=3.0.0,<4.0.0
pypdf>=5.0.0,<6.0.0
docx2txt>=0.8,<1.0
python-dotenv>=1.0.0,<2.0.0
numpy>=1.26.0,<3.0.0
